split Warehouse and Shipment Scene into classes
