+++ src/com/generic/utils/Persistent.java
@@ -4,10 +4,12 @@ import java.io.File;
 import java.io.FileNotFoundException;
 import java.io.FileReader;
 import java.io.IOException;
+
 import org.json.simple.JSONArray;
 import org.json.simple.JSONObject;
 import org.json.simple.parser.JSONParser;
 import org.json.simple.parser.ParseException;
+
 import com.generic.models.FreightType;
 import com.generic.models.Shipment;
 import com.generic.models.Warehouse;
@@ -17,14 +19,14 @@ public class Persistent {
 
 	// static ?
 	public Persistent() {}
-	
+
 	/**
 	 * Reads a file that is in JSON format containing various shipment information.
-	 * 
+	 *
 	 * @param filepath the path of JSON file
 	 */
 	@SuppressWarnings("unchecked")
-	public static void parseJson(String filepath) throws IOException, FileNotFoundException, ParseException {
+	public void parseJson(String filepath) throws IOException, FileNotFoundException, ParseException {
 		JSONParser jsonParser = new JSONParser();
 		FileReader reader = new FileReader(filepath);
 		JSONObject jsonFile = (JSONObject) jsonParser.parse(reader);
@@ -33,7 +35,7 @@ public class Persistent {
 
 		reader.close();
 	}
-	
+
 	/**
 	 * Reads a file that is in JSON format containing various shipment information.
 	 * @param JSON file object
@@ -55,14 +57,14 @@ public class Persistent {
 	 */
 	private static void parseWarehouseContentsToObjects(JSONObject shipmentObject) {
 		WarehouseTracker warehouseTracker = WarehouseTracker.getInstance();
-	
+
 		String warehouseID = (String) shipmentObject.get("warehouse_id");
 		// create warehouse
 		Warehouse warehouse = new Warehouse(warehouseID);
 		// build a shipment
 		Shipment shipment = new Shipment.Builder()
 				.id((String) shipmentObject.get("shipment_id"))
-				.type(FreightType.valueOf((String) shipmentObject.get("shipment_method").toString().toUpperCase()))
+				.type(FreightType.valueOf(shipmentObject.get("shipment_method").toString().toUpperCase()))
 				.weight(((Number) shipmentObject.get("weight")).doubleValue())
 				.date(((Number)shipmentObject.get("receipt_date")).longValue()).build();
 		// add the warehouse
@@ -70,8 +72,8 @@ public class Persistent {
 		// add the shipment to the warehouse
 		warehouseTracker.addShipment(warehouseID, shipment);
 	}
-	
-	
-	
+
+
+
 
 }
